fix: move component delegate-selector-drawer from apps/design-system to packages/ui/views
