+++ packages/ui/src/views/delegates/index.ts
@@ -2,3 +2,4 @@ export * from './types'
 export * from './components/delegate-connectivity-list'
 export * from './delegate-selector/delegate-selector-form'
 export * from './delegate-selector/delegate-selector-input'
+export * from './delegate-selector-drawer/delegate-selector-drawer'
